+++ ui/src/Pages/Preferences/Appearance/Index.scss
