--- ui/src/Pages/Preferences/FileBrowser.jsx
@@ -1,11 +1,0 @@
-function FileBrowser() {
-  return (
-    <section>
-      <h3>This section will house the following features:</h3>
-      <p>Change default path</p>
-      <p>Change to list or grid view</p>
-    </section>
-  );
-}
-
-export default FileBrowser;
