+++ ui/src/Pages/Preferences/Invites/Index.scss
