+++ ui/src/Pages/Preferences/Player/Index.scss
